# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Park Pal - Reserva Fácil API",
        "description": "API para sistema de reservas de estacionamiento",
        "contact": {
            "name": "Park Pal Support",
            "email": "admin@parkpal.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/vehiculos": {
            "get": {
                "tags": [
                    "Vehiculos"
                ],
                "summary": "Obtener lista de vehículos",
                "operationId": "b912521b0e41dec60b1fd4d98a0e06fe",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Término de búsqueda",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "ABC123"
                        }
                    },
                    {
                        "name": "tipo",
                        "in": "query",
                        "description": "Filtrar por tipo de vehículo",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "auto",
                                "motocicleta",
                                "camioneta"
                            ]
                        }
                    },
                    {
                        "name": "estado",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "activo",
                                "inactivo"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de vehículos obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Prohibido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Prohibido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vehiculos"
                ],
                "summary": "Crear un nuevo vehículo",
                "operationId": "20a2f18df869f0d5e7d5ddce18285c29",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patente",
                                    "marca",
                                    "modelo",
                                    "tipo",
                                    "color"
                                ],
                                "properties": {
                                    "patente": {
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "marca": {
                                        "type": "string",
                                        "example": "Toyota"
                                    },
                                    "modelo": {
                                        "type": "string",
                                        "example": "Corolla"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "enum": [
                                            "auto",
                                            "motocicleta",
                                            "camioneta"
                                        ],
                                        "example": "auto"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "Blanco"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 2022
                                    },
                                    "observaciones": {
                                        "type": "string",
                                        "example": "Vehículo en excelente estado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vehículo creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Prohibido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Prohibido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Usar el token JWT obtenido del login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}